feat: add Exit Rights section in AppPermissions
