+++ src/components/App.jsx
@@ -16,7 +16,7 @@ import Devices from 'containers/Devices'
 import IntentRedirect from 'services/IntentRedirect'
 import Passphrase from 'containers/Passphrase'
 import Permission from 'components/Permissions/PermissionDetails'
-import PermissionsApplication from 'components/Permissions/AppPermissions'
+import PermissionsApplication from 'components/Permissions/AppPermissions/AppPermissions'
 import PermissionsTab from 'components/Permissions/PermissionsTab'
 import DataPermissions from 'components/Permissions/DataPermissions/DataPermissions'
 import Profile from 'containers/Profile'
