+++ src/components/Permissions/helpers/permissionsHelper.js
@@ -44,3 +44,65 @@ export const getPermissionsVerbsByType = (permissions, type) => {
   }
   return verbs
 }
+
+export const isNotLastItem = (itemId, list) => {
+  if (list[0].name) {
+    return itemId !== list[list.length - 1].name
+  } else {
+    return itemId !== list[list.length - 1].id
+  }
+}
+
+export const filterPermissions = (remoteDoctypes, matchingQueryResult) => {
+  let limitedRightAccess = []
+  let exitRights = []
+  if (remoteDoctypes && matchingQueryResult?.data?.permissions) {
+    const permissions = Object.entries(matchingQueryResult.data.permissions)
+    for (let i = 0; i < permissions.length; i++) {
+      if (remoteDoctypes.includes(permissions[i][1].type)) {
+        exitRights.push(permissions[i])
+      } else {
+        limitedRightAccess.push(permissions[i])
+      }
+    }
+  }
+  return { limitedRightAccess, exitRights }
+}
+
+export const sortPermissionsByName = (completePermission, permissions, t) => {
+  return permissions
+    .map(([name, value]) => {
+      const type = value.type
+      const perm = t('CozyPermissions.Permissions.' + type)
+      return completePermission(name, perm, value, type)
+    })
+    .sort((a, b) => {
+      return a.title.localeCompare(b.title)
+    })
+}
+
+export const completeAppPermission = (
+  name,
+  permission,
+  { description, verbs },
+  type
+) => {
+  const completedPermission = {
+    name,
+    type,
+    title: permission
+  }
+  if (description) {
+    completedPermission.description = description
+  }
+  if (verbs) {
+    completedPermission.verbs = verbs
+  }
+  return completedPermission
+}
+
+export const sortDataByDate = queryResult => {
+  return queryResult.data.sort(
+    (a, b) => new Date(b.created_at) - new Date(a.created_at)
+  )
+}
