+++ src/locales/fr.json
@@ -455,7 +455,8 @@
     "limited_right_access": "Droit d'accès limité",
     "latest_data_releases": "Dernières sorties de données",
     "monthly_statistics": "Statistiques mensuelles",
-    "monthly_stats_phrase": "Compte rendu de la sortie de données depuis votre cozy vers %{app} le %{date}:"
+    "monthly_stats_phrase": "Compte rendu de la sortie de données depuis votre cozy vers %{app} le %{date}:",
+    "exit_rights": "Droits de sortie"
   },
   "Accessibility": {
     "previous": "Précédent"
